// packages/contexts/ingestion/src/blockchains/polkadot/types.ts
export interface TaostatsTransaction {
  amount: string;
  block: number;
  block_hash: string;
  block_number: number;
  confirmations: number;
  fee?: string;
  from: string;
  hash: string;
  success: boolean;
  timestamp: number;
  to: string;
}